requirements.txt
openai
python-dotenv
pymupdf
python-docx
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
pandas